+++ cpp00/ex01/contact.hpp
@@ -6,11 +6,11 @@
 class	Contact
 {
 	private:
-		std::string Name;
-		std::string Surname;
-		std::string	Nickname;
-		int			PhoneNumber;
-		std::string DarkSecret;
+		std::string 	Name;
+		std::string 	Surname;
+		std::string		Nickname;
+		long long int	PhoneNumber;
+		std::string 	DarkSecret;
 	public:
 		Contact()
 		{
@@ -20,17 +20,16 @@ class	Contact
 			PhoneNumber = 0;
 			DarkSecret = "";
 		}
-	int			get_phone_number(void);
-	int			ctrl_phone_number(std::string PhoneNumber);
-	void		set_name(std::string name);
-	void		set_surname(std::string surname);
-	void		set_nickname(std::string nickname);
-	void		set_phone_number(int phonenumber);
-	void		set_dark_secret(std::string darksecret);
-	std::string	get_name(void);
-	std::string	get_surname(void);
-	std::string	get_nickname(void);
-	std::string	get_dark_secret(void);
-
+	long long int	get_phone_number(void);
+	int				ctrl_phone_number(std::string PhoneNumber);
+	void			set_name(std::string name);
+	void			set_surname(std::string surname);
+	void			set_nickname(std::string nickname);
+	void			set_phone_number(long long int phonenumber);
+	void			set_dark_secret(std::string darksecret);
+	std::string		get_name(void);
+	std::string		get_surname(void);
+	std::string		get_nickname(void);
+	std::string		get_dark_secret(void);
 };
 #endif
